refactor(ae): split ae.c into cohesive TUs and build them per-object (#1221)
ae.c was 8,514 lines in one translation unit compiled by a single gcc
invocation, so any one-character edit recompiled the whole driver, the
dominant cost in the edit-ae-and-rebuild loop the cross-compile work
(#1208/#1216/#1218/#1220) touches constantly.

Two halves, both needed for the win the issue asks for:

1. Three command clusters move into their own sources reached through a
   new tools/ae_internal.h (the Toolchain struct plus the driver helpers
   they share):
     ae_cross.c    zig cross-compile backend (the hottest edit cluster)
     ae_version.c  version manager + the macOS Gatekeeper re-sign helper
     ae_repl.c     interactive REPL
   ae.c drops to 6,842 lines. macos_prepare_binary travels with the
   version manager (install re-signs a downloaded binary) but is also
   called from the build path, so it is declared in ae_internal.h.

2. The tools build becomes per-object. It was `gcc tools/*.c -o ae` in
   one shot, so splitting the sources alone would not have changed
   incremental cost. Each ae_*.c now compiles to its own object under a
   static pattern rule (so it wins over the generic $(OBJ_DIR)/%.o rule,
   which carries the wrong include path and defines), with -MMD -MP; the
   .d files are added to DEPS. Editing ae_cross.c now recompiles only
   ae_cross.o and relinks; editing ae_internal.h rebuilds exactly the
   four units that include it.

Pure code motion plus a build-graph change: the linked binary is
identical, ae spends its wall-clock in zig cc / posix_run / disk not its
own code so there is no runtime cost, and all four TUs compile clean
under clang and MinGW GCC's -Wall -Wextra -Werror. Verified: touch
ae_cross.c rebuilds one object; every command surface (run, build,
version, repl, cross) still works.
