+++ tools/ae.c
@@ -77,7 +77,7 @@ extern char** environ;
 // --------------------------------------------------------------------------
 // Cross-platform temp directory
 // --------------------------------------------------------------------------
-static const char* get_temp_dir(void) {
+const char* get_temp_dir(void) {
 #ifdef _WIN32
     const char* t = getenv("TEMP");
     if (!t) t = getenv("TMP");
@@ -94,33 +94,9 @@ static const char* get_temp_dir(void) {
 // Toolchain state
 // --------------------------------------------------------------------------
 
-typedef struct {
-    char root[1024];           // Aether root directory
-    char compiler[2048];       // Path to aetherc (root + /bin/aetherc = up to 1036 bytes)
-    char lib[1024];            // Path to libaether.a (if exists)
-    char include_flags[16384]; // -I flags for GCC. Sized to comfortably
-                               // hold the runtime/ + std/ + contrib/
-                               // sub-tree walks (issue #334 added contrib/
-                               // and the previous 4096 ceiling started
-                               // dropping dirs at the tail end of the walk).
-    char runtime_srcs[8192];   // Runtime .c files (source fallback)
-    bool has_lib;              // Whether precompiled lib exists
-    bool dev_mode;             // Running from source tree
-    bool verbose;              // Verbose output
-    /* Lib-search path forwarded to aetherc as one `--lib <dir>` flag
-     * per entry. Stored as an array (rather than a separator-string
-     * buffer) so we never re-construct a `dir1:dir2:dir3` string that
-     * has to survive shell quoting through system() — cmd.exe and
-     * MSYS2 between them mangle `;`-separated quoted strings unevenly,
-     * and one-flag-per-entry sidesteps the entire surface. Each
-     * `--lib X` from the user is parsed: if `X` is itself a separator-
-     * string, each piece is appended; if it's a single directory, it's
-     * appended verbatim. Issue #413. */
-    char lib_dirs[AETHER_LIB_DIRS_MAX][256];
-    int  lib_dir_count;
-} Toolchain;
+#include "ae_internal.h"
 
-static Toolchain tc = {0};
+Toolchain tc = {0};
 
 /* Append a directory (or a separator-string of directories) to
  * `tc.lib_dirs`. Used by every `--lib <X>` flag site so that
@@ -292,7 +268,7 @@ typedef struct {
 static bool g_coverage = false;
 
 // Build an aetherc command string with optional --lib flag
-static void build_aetherc_cmd(char* cmd, size_t cmd_size, const char* input, const char* output) {
+void build_aetherc_cmd(char* cmd, size_t cmd_size, const char* input, const char* output) {
     const char* emit_flag = "";
     if (g_emit_csrc)                   emit_flag = " --emit=csrc";
     else if (g_emit_lib && g_emit_exe) emit_flag = " --emit=both";
@@ -356,14 +332,14 @@ static void build_aetherc_cmd(char* cmd, size_t cmd_size, const char* input, con
 // Cache infrastructure
 // --------------------------------------------------------------------------
 
-static void mkdirs(const char* path);  // forward declaration
+
 
 static char s_cache_dir[512] = "";
 
 // Portable home-directory lookup.
 // On Windows: USERPROFILE (native shell) → HOME (MSYS2) → fallback.
 // On POSIX:   HOME → /tmp fallback.
-static const char* get_home_dir(void) {
+const char* get_home_dir(void) {
 #ifdef _WIN32
     const char* h = getenv("USERPROFILE");
     if (!h || !h[0]) h = getenv("HOME");
@@ -887,7 +863,7 @@ static int win_run(const char* cmd_str, int quiet) {
 }
 #endif
 
-static int run_cmd(const char* cmd) {
+int run_cmd(const char* cmd) {
 #ifndef _WIN32
     return posix_run(cmd, 0);
 #else
@@ -896,7 +872,7 @@ static int run_cmd(const char* cmd) {
 }
 
 // Run a command, suppressing all output (quiet mode)
-static int run_cmd_quiet(const char* cmd) {
+int run_cmd_quiet(const char* cmd) {
 #ifndef _WIN32
     return posix_run(cmd, 1);
 #else
@@ -905,7 +881,7 @@ static int run_cmd_quiet(const char* cmd) {
 }
 
 // Run a command, showing stderr (warnings) but hiding stdout
-static int run_cmd_show_warnings(const char* cmd) {
+int run_cmd_show_warnings(const char* cmd) {
 #ifndef _WIN32
     return posix_run(cmd, 2);
 #else
@@ -960,9 +936,8 @@ static int append_manifest_srcs(char* out, size_t out_sz,
     return pos > 0;
 }
 
-static void macos_prepare_binary(const char* path);
 
-static bool path_exists(const char* path) {
+bool path_exists(const char* path) {
 #ifdef _WIN32
     DWORD attrs = GetFileAttributesA(path);
     return (attrs != INVALID_FILE_ATTRIBUTES && !(attrs & FILE_ATTRIBUTE_DIRECTORY));
@@ -986,12 +961,12 @@ static bool is_safe_shell_arg(const char* s) {
     return true;
 }
 
-static bool dir_exists(const char* path) {
+bool dir_exists(const char* path) {
     struct stat st;
     return stat(path, &st) == 0 && S_ISDIR(st.st_mode);
 }
 
-static void mkdirs(const char* path) {
+void mkdirs(const char* path) {
     char tmp[1024];
     snprintf(tmp, sizeof(tmp), "%s", path);
     for (char* p = tmp + 1; *p; p++) {
@@ -1895,7 +1870,7 @@ fail:
 
 // Get cflags from aether.toml [build] section (applied only for release/ae-build)
 // Returns empty string if not found or no aether.toml
-static const char* get_cflags(void) {
+const char* get_cflags(void) {
     static char flags[512] = "";
     static bool checked = false;
 
@@ -2240,7 +2215,7 @@ static const char* opt_flags(bool optimize) {
     return optimize ? "-O2 -Wformat" : "-O0 -g -Wformat";
 }
 
-static void build_gcc_cmd(char* cmd, size_t size,
+void build_gcc_cmd(char* cmd, size_t size,
                           const char* c_file, const char* out_file,
                           bool optimize, const char* extra_files) {
     const char* link_flags = get_link_flags();
@@ -3463,7 +3438,7 @@ typedef struct {
 } CapturedFunction;
 
 /* Run aetherc with the given args and capture stdout into out_buf. */
-static int aetherc_capture_stdout(const char* arg1, const char* in_path,
+int aetherc_capture_stdout(const char* arg1, const char* in_path,
                                   const char* arg2_or_null,
                                   char* out_buf, size_t out_size) {
     char cmd[4096];
@@ -5052,460 +5027,6 @@ int cmd_build_namespace(int argc, char** argv) {
     return rc;
 }
 
-/* ------------------------------------------------------------------ *
- *  Cross-compilation via a zig cc backend (#1105)
- *
- *  `ae build --target=<triple>` builds a foreign-target binary using
- *  zig as a self-contained cross-compiler: zig bundles each target's
- *  libc, system headers and linker, so the Aether runtime and stdlib
- *  compile straight from source for the target. The platform backend
- *  (epoll vs kqueue, spawn_sandboxed_linux vs bsd) is chosen by the
- *  compile-time __linux__ / __APPLE__ macros zig predefines for the
- *  target, so one source set serves every target with no per-host
- *  file selection.
- *
- *  PR 1 scope: dependency-free programs. Stdlib modules whose C code
- *  needs an external library we cannot yet cross-build (openssl,
- *  nghttp2, zlib, pcre2) are left out of the compile set, and a
- *  program importing one is rejected up front with a clear message.
- *  Networking / crypto / compression / regex cross builds are the
- *  documented follow-up. Native builds are entirely unaffected.
- * ------------------------------------------------------------------ */
-
-/* Map an Aether target string to a zig `-target` triple. Returns NULL
- * for anything that isn't a supported cross triple (native / wasm /
- * unknown), which the caller treats as "not a cross build". */
-static const char* cross_target_to_zig(const char* t) {
-    if (!t) return NULL;
-    if (!strcmp(t, "aarch64-macos") || !strcmp(t, "arm64-macos"))  return "aarch64-macos-none";
-    if (!strcmp(t, "x86_64-macos")  || !strcmp(t, "amd64-macos"))  return "x86_64-macos-none";
-    if (!strcmp(t, "aarch64-linux") || !strcmp(t, "arm64-linux"))  return "aarch64-linux-gnu";
-    if (!strcmp(t, "x86_64-linux")  || !strcmp(t, "amd64-linux"))  return "x86_64-linux-gnu";
-    /* Windows (Tier A — self-contained): zig bundles the full MinGW-w64 target
-     * (CRT, Win32 headers, import libs), so no base sysroot, no --sysroot, no
-     * CRT/libc dance — identical to the linux/macos arms. The runtime's _WIN32
-     * guards (already exercised by native MSYS2 builds) compile against zig's
-     * mingw-w64 bundle. cross_target_needs_sysroot stays false for windows. */
-    if (!strcmp(t, "x86_64-windows")  || !strcmp(t, "amd64-windows")) return "x86_64-windows-gnu";
-    if (!strcmp(t, "aarch64-windows") || !strcmp(t, "arm64-windows")) return "aarch64-windows-gnu";
-    /* FreeBSD (Tier B): zig cc does NOT bundle a FreeBSD libc, so these
-     * additionally require a base sysroot (headers + libc) provided via
-     * AETHER_SYSROOT — see cross_target_needs_sysroot() and the
-     * aether-crossbuild repo (scripts/fetch-freebsd-base.sh). The zig target
-     * has no OS-version; the base sysroot carries FreeBSD 14 vs 15. */
-    if (!strcmp(t, "aarch64-freebsd") || !strcmp(t, "arm64-freebsd")) return "aarch64-freebsd";
-    if (!strcmp(t, "x86_64-freebsd")  || !strcmp(t, "amd64-freebsd")) return "x86_64-freebsd";
-    return NULL;
-}
-
-/* True if `t` is a cross target whose libc zig cc does NOT bundle, so a base
- * sysroot must be supplied (via AETHER_SYSROOT). Tier A (macos/linux) is
- * self-contained; Tier B (freebsd) is not. */
-static bool cross_target_needs_sysroot(const char* t) {
-    if (!t) return false;
-    return strstr(t, "freebsd") != NULL;
-}
-
-/* Locate the authoritative source MANIFEST and the base directory its
- * entries are relative to. Dev tree: <root>/build/MANIFEST (entries
- * relative to <root>). Installed: <root>/share/aether/MANIFEST
- * (entries relative to <root>/share/aether). Returns false if neither
- * exists. */
-static bool cross_find_manifest(char* manifest, size_t msz,
-                                char* base, size_t bsz) {
-    snprintf(manifest, msz, "%s/build/MANIFEST", tc.root);
-    if (path_exists(manifest)) { snprintf(base, bsz, "%s", tc.root); return true; }
-    snprintf(manifest, msz, "%s/share/aether/MANIFEST", tc.root);
-    if (path_exists(manifest)) { snprintf(base, bsz, "%s/share/aether", tc.root); return true; }
-    return false;
-}
-
-/* Read MANIFEST into `out`, one absolute source path per entry, for
- * every link-suitable source. Returns the count (capped at `max`), or
- * -1 on error (unreadable manifest). `zig cc -c` emits only one object
- * when handed several sources at once, so the caller compiles each path
- * individually.
- *
- * Every runtime/stdlib source compiles for a cross target with no
- * external library: each openssl / nghttp2 / zlib / pcre2 dependency is
- * behind an AETHER_HAS_* guard that falls to a graceful "unavailable"
- * stub when the macro is undefined (which it is here). So we build the
- * full set, archive it, and let the final link pull only the objects the
- * program references, exactly as a native `-laether` link against the
- * complete libaether.a does. Library-backed features (TLS, real crypto,
- * regex, zlib, HTTP/2) then report unavailable at runtime on the target,
- * exactly like a native build on a host without those libraries, while
- * pure helpers such as base64 (std.encoding) keep working. */
-#define CROSS_SRC_PATH_MAX 1024
-static int cross_collect_core_list(char out[][CROSS_SRC_PATH_MAX], int max,
-                                   const char* manifest, const char* base) {
-    FILE* f = fopen(manifest, "r");
-    if (!f) return -1;
-    int count = 0;
-    char line[1024];
-    while (fgets(line, sizeof(line), f) && count < max) {
-        size_t n = strlen(line);
-        while (n && (line[n-1] == '\n' || line[n-1] == '\r' ||
-                     line[n-1] == ' '  || line[n-1] == '\t')) line[--n] = '\0';
-        char* p = line;
-        while (*p == ' ' || *p == '\t') p++;
-        if (*p == '\0' || *p == '#') continue;                 /* comment / blank */
-        size_t plen = strlen(p);
-        if (plen < 2 || strcmp(p + plen - 2, ".c") != 0) continue;
-        int need = snprintf(out[count], CROSS_SRC_PATH_MAX, "%s/%s", base, p);
-        if (need < 0 || need >= CROSS_SRC_PATH_MAX) {
-            /* A truncated source path would compile the wrong file or
-             * fail obscurely at the C compiler; skip it loudly instead. */
-            fprintf(stderr, "Warning: source path too long, skipped: %s/%s\n", base, p);
-            continue;
-        }
-        count++;
-    }
-    fclose(f);
-    return count;
-}
-
-/* Scan a program's (transitive) imports via `aetherc --emit=inspect`.
- * If it uses a stdlib module with library-backed features that are not
- * cross-built (so they report "unavailable" at runtime on the target),
- * write the module name to `which` and return true. Used to warn, not
- * to block: the program still builds and the non-library parts still
- * work. On inspect failure returns false (no warning). */
-static bool cross_uses_unsupported_module(const char* file, char* which, size_t wsz) {
-    static char out[65536];
-    if (aetherc_capture_stdout("--emit=inspect", file, NULL, out, sizeof(out)) != 0)
-        return false;
-    static const char* mods[] = {
-        "std.http", "std.net", "std.cryptography", "std.regex", "std.zlib",
-        "std.encoding", NULL   /* base64 in std.encoding is openssl-backed */
-    };
-    for (int i = 0; mods[i]; i++) {
-        if (strstr(out, mods[i])) { snprintf(which, wsz, "%s", mods[i]); return true; }
-    }
-    return false;
-}
-
-/* Execute a full cross build: compile the dependency-free core to
- * per-file objects, archive them, then link the program against the
- * archive. Linking against an archive (rather than force-linking every
- * runtime object into the image) reproduces a native `-laether` link's
- * on-demand object pulling: an object is pulled only when the program
- * references one of its symbols. That is what lets a user program define
- * a top-level function named like an *unreferenced* runtime global
- * (e.g. `describe`, `notify`, `event` in aether_host.c) without a
- * duplicate-symbol clash, exactly as a native build allows. Returns 0
- * on success, non-zero (with a diagnostic) otherwise. POSIX host only.
- *
- * Feature defines mirror the runtime archive's normal Makefile CFLAGS:
- * AETHER_HAS_SANDBOX is the only one not auto-derived by
- * aether_optimization_config.h (filesystem / networking / threading
- * default on when no AETHER_NO_* is passed). The external-library macros
- * (AETHER_HAS_OPENSSL / _ZLIB / _NGHTTP2 / _PCRE2) are deliberately left
- * undefined so their stub paths compile, matching the excluded sources. */
-static int run_cross_build(const char* c_file, const char* out_file,
-                           bool optimize, const char* extra,
-                           const char* ztriple) {
-    char manifest[2048], base[1024];
-    if (!cross_find_manifest(manifest, sizeof(manifest), base, sizeof(base))) {
-        fprintf(stderr,
-            "Error: cross-compilation needs the runtime source MANIFEST, which was "
-            "not found (looked under %s/build and %s/share/aether). Run `make stdlib` "
-            "in a source tree, or reinstall the toolchain.\n", tc.root, tc.root);
-        return 1;
-    }
-    static char srcs[256][CROSS_SRC_PATH_MAX];
-    int n = cross_collect_core_list(srcs, 256, manifest, base);
-    if (n <= 0) {
-        fprintf(stderr, "Error: could not assemble the cross-compile source set from %s.\n",
-                manifest);
-        return 1;
-    }
-
-    /* Fresh per-build object directory under the system temp. */
-    char objdir[1024];
-    snprintf(objdir, sizeof(objdir), "%s/ae-cross-%d", get_temp_dir(), (int)getpid());
-    mkdirs(objdir);
-
-    const char* user_cflags = get_cflags();
-    const char* opt = optimize ? "-O2" : "-O0 -g";
-    const char* ex = extra ? extra : "";
-    /* std.audio's vendored miniaudio auto-selects a backend by platform macro:
-     * on a macos target it #includes <CoreAudio/CoreAudio.h>, an APPLE FRAMEWORK
-     * header that zig's bundled macOS SDK stubs deliberately do NOT ship (the
-     * Apple-licensed part). A cross-built binary can't use CoreAudio without a
-     * real Mac SDK anyway, so disable it: -DMA_NO_COREAUDIO makes miniaudio fall
-     * back to its null backend (std.audio reports unavailable at runtime — same
-     * warn-and-degrade as openssl/nghttp2 on a Tier-B target). Without this, a
-     * macos cross-build of ANY program fails compiling aether_audio.c, even one
-     * that never touches std.audio (the runtime always compiles it). */
-    char feature_defs[128] = "-DAETHER_HAS_SANDBOX";
-    if (strstr(ztriple, "macos")) {
-        strncat(feature_defs, " -DMA_NO_COREAUDIO",
-                sizeof(feature_defs) - strlen(feature_defs) - 1);
-    }
-    /* Tier-B (FreeBSD) targets need a base sysroot zig cc doesn't bundle;
-     * AETHER_SYSROOT points at it (bases/<cpu>-freebsd[ver]/ from
-     * aether-crossbuild). Applied to BOTH compile and link. Empty for the
-     * self-contained Tier-A targets. NB: for a FreeBSD target, `--sysroot`
-     * ALONE does not make zig cc search the sysroot's usr/include and
-     * usr/lib (unlike its bundled targets) — the -I/-L must be explicit
-     * (verified with zig 0.13). */
-    char sysroot_flag[3200];   /* COMPILE flags: --sysroot + -I/-L */
-    char fbsd_link[4096];      /* LINK tail: CRT objects + the real libc.so.7 */
-    char fbsd_platform_libs[2048]; /* LINK tail: FreeBSD platform -l names */
-    char win_platform_libs[512];   /* LINK tail: Windows system -l names */
-    sysroot_flag[0] = '\0';
-    fbsd_link[0] = '\0';
-    fbsd_platform_libs[0] = '\0';
-    win_platform_libs[0] = '\0';
-    /* Windows system libs. std.cryptography's OS RNG uses BCryptGenRandom
-     * (bcrypt.dll); winsock, crypt32, advapi32 etc. are pulled by the runtime
-     * and static openssl. Same shape as fbsd_platform_libs (casper) — a
-     * target-specific always-on set the cross link must append (zig bundles the
-     * mingw CRT but NOT these -l names). Matches the native Windows
-     * win_link_libs. Detected off the zig triple (…-windows-gnu). */
-    if (strstr(ztriple, "windows")) {
-        snprintf(win_platform_libs, sizeof(win_platform_libs),
-                 "-lws2_32 -lcrypt32 -lgdi32 -luser32 -ladvapi32 -lbcrypt -ldbghelp");
-    }
-    if (cross_target_needs_sysroot(ztriple)) {
-        const char* sr = getenv("AETHER_SYSROOT");
-        if (!sr || !*sr) {
-            fprintf(stderr,
-                "Error: target %s needs a FreeBSD base sysroot, but AETHER_SYSROOT is unset.\n"
-                "  zig cc does not bundle a FreeBSD libc. Provision one with aether-crossbuild:\n"
-                "    ./scripts/fetch-freebsd-base.sh <cpu> [major]   # e.g. x86_64 15\n"
-                "  then: AETHER_SYSROOT=<crossbuild>/bases/<cpu>-freebsd[ver] ae build ... --target=%s\n",
-                ztriple, ztriple);
-            return 1;
-        }
-        snprintf(sysroot_flag, sizeof(sysroot_flag),
-                 "--sysroot=%s -I%s/usr/include -L%s/usr/lib -L%s/lib",
-                 sr, sr, sr, sr);
-        /* zig cc can't provide a FreeBSD libc ("error: libc not available"),
-         * and the sysroot's usr/lib/libc.so is a GROUP linker script naming
-         * ABSOLUTE /lib paths that don't exist on this build host. So link
-         * FreeBSD explicitly: -nostdlib + the base's CRT startup objects
-         * (crt1/crti/crtn — crt1 pulls __libc_start1, a FreeBSD-15 symbol) +
-         * the real versioned libc.so.7. Verified end-to-end with zig 0.13
-         * against a FreeBSD-15 base sysroot. */
-        /* libthr.so.3 (POSIX threads) BY EXPLICIT PATH, next to libc.so.7. The
-         * Aether runtime (scheduler, actor threads) and std.http's server call
-         * pthread_create/mutex/cond, so a FreeBSD cross-link fails `undefined
-         * symbol: pthread_create` regardless of what the app imports (the native
-         * FreeBSD build gets this via -pthread, ae.c ~2367). It MUST be a path,
-         * not `-lpthread`: base/usr/lib/libpthread.so is a symlink onto libthr,
-         * and -lthr/-lpthread does NOT resolve under zig-lld + -nostdlib against
-         * this split base (verified: the -l form leaves pthread_create undefined,
-         * the explicit libthr.so.3 path links clean). Versioned .so.3 like
-         * libc.so.7 above. */
-        snprintf(fbsd_link, sizeof(fbsd_link),
-                 "-nostdlib \"%s/usr/lib/crt1.o\" \"%s/usr/lib/crti.o\" "
-                 "\"%s/lib/libc.so.7\" \"%s/lib/libthr.so.3\" \"%s/usr/lib/crtn.o\"",
-                 sr, sr, sr, sr, sr);
-
-        /* Platform libs the FreeBSD link needs, mirroring the NATIVE FreeBSD
-         * build (ae.c ~2368). The base sysroot's -L (from sysroot_flag) already
-         * resolves these; only the -l names were missing on the cross path.
-         *
-         * casper — ALWAYS: std/casper/aether_casper.c is unconditionally in
-         * libaether.a and calls cap_getpwnam / cap_sysctlbyname / cap_getaddrinfo
-         * / ..., so a FreeBSD cross-link fails with `undefined symbol: cap_*`
-         * regardless of what the app imports. The base ships all of them. The
-         * complete set, per the cap_* symbols aether_casper.c references:
-         *   cap_init/cap_close/cap_service_open -> libcasper (core)
-         *   cap_getpwnam                        -> libcap_pwd
-         *   cap_sysctl(byname)                  -> libcap_sysctl
-         *   (grp)                               -> libcap_grp
-         *   cap_getaddrinfo (aether_casper_resolve, DNS) -> libcap_dns
-         * cap_dns was the one #1216 missed (its DNS path only surfaces after
-         * pwd/sysctl resolve). We emit the names literally rather than via
-         * AETHER_CASPER_LIBS — that macro is populated by globbing the HOST's
-         * /lib when `ae` is built on FreeBSD, and is empty in an `ae`
-         * cross-compiled/built on Linux.
-         *
-         * casper is FreeBSD-only. The openssl/nghttp2/zlib/pcre2 (Tier-2) libs
-         * are target-AGNOSTIC and handled by crossbuild_libs below, so they
-         * work for windows/linux/macos too — not just freebsd. */
-        snprintf(fbsd_platform_libs, sizeof(fbsd_platform_libs),
-                 "-lcasper -lcap_pwd -lcap_sysctl -lcap_grp -lcap_dns");
-    }
-
-    /* Tier-2 libs from a CROSSBUILD_SYSROOT — CONDITIONAL and TARGET-AGNOSTIC.
-     * openssl/nghttp2/zlib/pcre2 are not bundled by zig for any target;
-     * aether-crossbuild's provision.sh <triple> builds them into
-     * sysroots/<triple>/. When that sysroot is provided (same CROSSBUILD_SYSROOT
-     * contract #1213 gave contrib_build.sh), append its -L + the -l names so
-     * std.cryptography / std.http / std.zlib / std.regex link for real — for
-     * FreeBSD AND Windows AND linux/macos. Without it, warn-and-omit stands
-     * (the features report unavailable at runtime). The -l names are the same
-     * across targets; only the -L (the sysroot) differs. */
-    char crossbuild_libs[2048];
-    crossbuild_libs[0] = '\0';
-    {
-        const char* xsr = getenv("CROSSBUILD_SYSROOT");
-        if (xsr && *xsr) {
-            /* Append each -l ONLY when that lib is actually staged in the
-             * sysroot. provision.sh may build a subset (a target might have
-             * pcre2 + zlib but not openssl yet), and zig hard-errors on a
-             * requested-but-absent lib. Probing per-lib links exactly what's
-             * provisioned — the same "link what's there" discipline #1213's
-             * contrib cross mode uses. openssl is -lssl + -lcrypto (both from
-             * libssl.a/libcrypto.a); the rest are 1:1. */
-            size_t p = 0;
-            p += (size_t)snprintf(crossbuild_libs + p, sizeof(crossbuild_libs) - p,
-                                  "-L%s/lib", xsr);
-            char probe[2600];
-            struct { const char* lib; const char* names; } t2[] = {
-                { "ssl",     "-lssl -lcrypto" },  /* libssl.a present => both */
-                { "nghttp2", "-lnghttp2" },
-                { "z",       "-lz" },
-                { "pcre2-8", "-lpcre2-8" },
-                /* contrib.sqlite: the Aether veneer archive
-                 * (libaether_sqlite.a, from contrib_build.sh CONTRIB_TARGET
-                 * mode) BEFORE the underlying C lib (libsqlite3.a) — ld.lld's
-                 * single pass needs the veneer's sqlite3_* references resolved
-                 * by the lib that follows. Probed on the VENEER so a program
-                 * that doesn't use sqlite links nothing extra. */
-                { "aether_sqlite", "-laether_sqlite -lsqlite3" },
-                /* contrib.host.python: the embedded-Python bridge veneer. NO
-                 * -lpython — the bridge dlopen()s the deploy host's libpython
-                 * at runtime (AETHER_PYTHON_SONAME), so the .a has no
-                 * unresolved CPython symbols and needs no python at link. Just
-                 * the veneer archive. Probed on it so a program not embedding
-                 * python links nothing extra. */
-                { "aether_host_python", "-laether_host_python" },
-                /* contrib.host.ruby: same dlopen model as python (no -lruby;
-                 * the deploy host's libruby is dlopen'd at runtime). Veneer
-                 * only. */
-                { "aether_host_ruby", "-laether_host_ruby" },
-            };
-            for (size_t i = 0; i < sizeof(t2) / sizeof(t2[0]); i++) {
-                snprintf(probe, sizeof(probe), "%s/lib/lib%s.a", xsr, t2[i].lib);
-                if (path_exists(probe) && p < sizeof(crossbuild_libs)) {
-                    p += (size_t)snprintf(crossbuild_libs + p, sizeof(crossbuild_libs) - p,
-                                          " %s", t2[i].names);
-                }
-            }
-        }
-    }
-    static char cmd[24576];
-    /* Accumulated quoted "<objpath>" list, in compile order, for the ar
-     * step. posix_run tokenizes the command itself (no shell), so the
-     * archive must name each object explicitly rather than glob. */
-    static char objlist[24576];
-    size_t obj_pos = 0;
-    objlist[0] = '\0';
-    int rc = 1;
-
-    int w;
-    do {
-        /* 1. Compile each core source to its own object in objdir. zig cc
-         *    -c emits only one object when handed several sources at once,
-         *    so compile one at a time (all core basenames are unique). */
-        bool compile_failed = false;
-        for (int i = 0; i < n; i++) {
-            const char* bn = strrchr(srcs[i], '/');
-            bn = bn ? bn + 1 : srcs[i];
-            char objpath[2048];
-            /* basename with its trailing ".c" replaced by ".o" */
-            snprintf(objpath, sizeof(objpath), "%s/%.*so", objdir,
-                     (int)(strlen(bn) - 1), bn);
-            w = snprintf(cmd, sizeof(cmd),
-                "zig cc -target %s %s %s %s %s %s -c \"%s\" -o \"%s\"",
-                ztriple, sysroot_flag, opt, feature_defs, user_cflags, tc.include_flags,
-                srcs[i], objpath);
-            if (w < 0 || (size_t)w >= sizeof(cmd)) {
-                fprintf(stderr, "Error: cross-compile command exceeded the %zu-byte buffer.\n",
-                        sizeof(cmd));
-                compile_failed = true;
-                break;
-            }
-            if (run_cmd_show_warnings(cmd) != 0) {
-                fprintf(stderr, "Error: cross-compiling %s for %s failed.\n", srcs[i], ztriple);
-                compile_failed = true;
-                break;
-            }
-            int ow = snprintf(objlist + obj_pos, sizeof(objlist) - obj_pos,
-                              "%s\"%s\"", obj_pos ? " " : "", objpath);
-            if (ow < 0 || obj_pos + (size_t)ow >= sizeof(objlist)) {
-                fprintf(stderr, "Error: cross-compile object list overflowed its buffer.\n");
-                compile_failed = true;
-                break;
-            }
-            obj_pos += (size_t)ow;
-        }
-        if (compile_failed) break;
-
-        /* 2. Archive the objects (named explicitly, no glob) so the final
-         *    link pulls only what the program references (native
-         *    `-laether` semantics). */
-        w = snprintf(cmd, sizeof(cmd),
-            "zig ar rcs \"%s/libaether.a\" %s", objdir, objlist);
-        if (w < 0 || (size_t)w >= sizeof(cmd)) {
-            fprintf(stderr, "Error: cross-compile archive command exceeded the %zu-byte buffer.\n",
-                    sizeof(cmd));
-            break;
-        }
-        if (run_cmd_show_warnings(cmd) != 0) {
-            fprintf(stderr, "Error: archiving the cross runtime failed.\n");
-            break;
-        }
-
-        /* Clear any stale output so the FreeBSD "output exists == linked"
-         * success signal below can't be fooled by a prior build's binary. */
-        remove(out_file);
-
-        /* 3. Link the program against the archive. FreeBSD (Tier B) needs the
-         *    explicit CRT objects + real libc.so.7 (fbsd_link); zig's bundled
-         *    FreeBSD-14 libc can't satisfy a 15 base's __libc_start1. The CRT
-         *    objects bracket the program/runtime; libm comes from the sysroot
-         *    -L. Tier A keeps the compact -lm form. */
-        if (fbsd_link[0]) {
-            /* Platform -l names go AFTER libaether.a — it references their
-             * symbols (casper's cap_*, openssl's SSL_*, …), so they must
-             * follow it on the link line for ld.lld's single-pass resolution. */
-            w = snprintf(cmd, sizeof(cmd),
-                "zig cc -target %s %s %s %s %s %s \"%s\" %s \"%s/libaether.a\" %s %s -lm -o \"%s\"",
-                ztriple, sysroot_flag, fbsd_link, opt, feature_defs, tc.include_flags,
-                c_file, ex, objdir, fbsd_platform_libs, crossbuild_libs, out_file);
-        } else {
-            /* Tier A (linux/macos/windows): compact form + any CROSSBUILD_SYSROOT
-             * Tier-2 libs AND the Windows system libs after libaether.a (it
-             * references their symbols — BCryptGenRandom etc.). */
-            w = snprintf(cmd, sizeof(cmd),
-                "zig cc -target %s %s %s %s %s \"%s\" %s \"%s/libaether.a\" %s %s -o \"%s\" -lm",
-                ztriple, sysroot_flag, opt, feature_defs, tc.include_flags, c_file, ex, objdir, crossbuild_libs, win_platform_libs, out_file);
-        }
-        if (w < 0 || (size_t)w >= sizeof(cmd)) {
-            fprintf(stderr, "Error: cross-compile link command exceeded the %zu-byte buffer.\n",
-                    sizeof(cmd));
-            break;
-        }
-        if (run_cmd_show_warnings(cmd) != 0) {
-            /* zig cc exits nonzero on a FreeBSD -nostdlib link with a COSMETIC
-             * "error: libc not available" note, even though clang+lld produced
-             * a valid binary (zig reserves that message for targets it can't
-             * supply a libc for — which is exactly why we bring our own). A
-             * GENUINE link failure (undefined symbol) leaves NO output file, so
-             * "the output exists and is non-empty" cleanly separates the two.
-             * Verified with zig 0.13 against a FreeBSD-15 base sysroot. */
-            if (!(fbsd_link[0] && path_exists(out_file))) {
-                fprintf(stderr, "Error: cross-linking for %s failed.\n", ztriple);
-                break;
-            }
-            fprintf(stderr, "note: zig cc reported \"libc not available\" for %s; "
-                            "the FreeBSD binary linked correctly regardless.\n", ztriple);
-        }
-        rc = 0;
-    } while (0);
-
-    /* Best-effort removal of the temp object tree. */
-    char rmcmd[1100];
-    snprintf(rmcmd, sizeof(rmcmd), "rm -rf \"%s\"", objdir);
-    (void)system(rmcmd);
-    return rc;
-}
 
 static int cmd_build(int argc, char** argv) {
     const char* file = NULL;
@@ -6600,1198 +6121,6 @@ static int cmd_examples(int argc, char** argv) {
 // REPL session: accumulated lines that persist across evaluations.
 // Each entry is a statement (assignment, function def, etc.) that gets
 // replayed before the current input so variables/functions stay in scope.
-#define REPL_MAX_LINES 256
-#define REPL_LINE_LEN  1024
-
-// Compile and run the REPL input. Returns 1 on success, 0 on failure.
-static int repl_eval(const char* ae_file, const char* c_file,
-                     const char* exe_file, char** history,
-                     int history_count, const char* input) {
-    FILE* f = fopen(ae_file, "w");
-    if (!f) return 0;
-    fprintf(f, "main() {\n");
-    for (int i = 0; i < history_count; i++)
-        fprintf(f, "    %s\n", history[i]);
-    const char* rest = input;
-    const char* nl;
-    while ((nl = strchr(rest, '\n')) != NULL) {
-        fprintf(f, "    %.*s\n", (int)(nl - rest), rest);
-        rest = nl + 1;
-    }
-    if (*rest) fprintf(f, "    %s\n", rest);
-    fprintf(f, "}\n");
-    fclose(f);
-
-    char cmd[16384];
-    build_aetherc_cmd(cmd, sizeof(cmd), ae_file, c_file);
-    if (run_cmd_quiet(cmd) != 0) {
-        run_cmd(cmd);
-        remove(c_file);
-        return 0;
-    }
-    build_gcc_cmd(cmd, sizeof(cmd), c_file, exe_file, false, NULL);
-    if (run_cmd_quiet(cmd) != 0) {
-        build_gcc_cmd(cmd, sizeof(cmd), c_file, exe_file, false, NULL);
-        run_cmd(cmd);
-        remove(c_file);
-        remove(exe_file);
-        return 0;
-    }
-    snprintf(cmd, sizeof(cmd), "\"%s\"", exe_file);
-    /* Force any pending parent-side stdio to drain BEFORE the child
-     * inherits the fd. Without this, prompts that the parent printed
-     * but hadn't yet flushed sit in the userspace FILE* buffer; the
-     * child writes its output directly to the kernel-side fd; then
-     * when the parent eventually flushes, its delayed prompt
-     * overwrites or interleaves with the child's output in the pipe
-     * stream. Under parallel CI load on Linux, this race becomes
-     * deterministic and the child's output appears lost.
-     *
-     * The fflush-before pattern is well-established stdio hygiene
-     * around any fork/exec/posix_spawn that shares stdout with the
-     * parent's libc-buffered stream. Same hygiene applied after the
-     * child runs so subsequent parent prompts arrive in the right
-     * order without a second eval. */
-    fflush(stdout);
-    fflush(stderr);
-    run_cmd(cmd);
-    fflush(stdout);
-    fflush(stderr);
-    remove(c_file);
-    remove(exe_file);
-    return 1;
-}
-
-// Persist an assignment or const into session history, replacing
-// previous assignments to the same variable name.
-static void repl_persist(char** history, int* history_count, const char* input) {
-    char* eq = strchr(input, '=');
-    int has_assign = (eq && (eq == input ||
-        (eq[-1] != '=' && eq[-1] != '!' && eq[-1] != '<' && eq[-1] != '>'))
-        && eq[1] != '=');
-    int has_const = (strncmp(input, "const ", 6) == 0);
-    if (!has_assign && !has_const) return;
-
-    int replaced = 0;
-    if (has_assign && eq) {
-        int name_len = (int)(eq - input);
-        while (name_len > 0 && input[name_len - 1] == ' ') name_len--;
-        for (int i = 0; i < *history_count; i++) {
-            char* heq = strchr(history[i], '=');
-            if (heq) {
-                int hlen = (int)(heq - history[i]);
-                while (hlen > 0 && history[i][hlen - 1] == ' ') hlen--;
-                if (hlen == name_len && strncmp(input, history[i], name_len) == 0) {
-                    free(history[i]);
-                    history[i] = strdup(input);
-                    replaced = 1;
-                    break;
-                }
-            }
-        }
-    }
-    if (!replaced && *history_count < REPL_MAX_LINES)
-        history[(*history_count)++] = strdup(input);
-}
-
-// Check if a single line is a complete statement (no open braces).
-// Single-line statements execute immediately without waiting for blank line.
-static int repl_is_complete_line(const char* line) {
-    int depth = 0;
-    for (const char* p = line; *p; p++) {
-        if (*p == '{') depth++;
-        else if (*p == '}') depth--;
-    }
-    return depth == 0;
-}
-
-static int cmd_repl(void) {
-    /* When the REPL's stdin is a pipe (CI tests, scripted invocations),
-     * glibc defaults stdout to FULLY-buffered. Prompts that the user
-     * needs to see immediately get held in the userspace buffer until
-     * the next fflush or program exit — and any child process the
-     * REPL spawns (`repl_eval` invokes aetherc + gcc + the compiled
-     * binary) writes DIRECTLY to the kernel fd, bypassing the
-     * parent's buffer. The result is interleaved or lost output in
-     * the pipe stream. Switching to line buffering makes every
-     * newline-terminated write go straight to the fd, which is what
-     * an interactive shell user gets on a TTY anyway. Same fix for
-     * stderr in case error messages arrive while the parent has
-     * pending stdout. No-op on a real TTY (already line-buffered). */
-    setvbuf(stdout, NULL, _IOLBF, BUFSIZ);
-    setvbuf(stderr, NULL, _IOLBF, BUFSIZ);
-
-    printf("\n");
-    // Dynamic box: "   Aether X.Y.Z REPL   "
-    int ver_len = (int)strlen(AE_VERSION);
-    int title_len = 15 + ver_len;  // "   Aether " (10) + ver + " REPL" (5)
-    int help_len  = 21;            // "   :help for commands"
-    int inner = title_len + 3;     // 3 chars right padding
-    if (inner < help_len + 3) inner = help_len + 3;
-    printf("  ┌"); for (int i = 0; i < inner; i++) printf("─"); printf("┐\n");
-    printf("  │   Aether %s REPL", AE_VERSION);
-    for (int i = title_len; i < inner; i++) printf(" ");
-    printf("│\n");
-    printf("  │   :help for commands");
-    for (int i = help_len; i < inner; i++) printf(" ");
-    printf("│\n");
-    printf("  └"); for (int i = 0; i < inner; i++) printf("─"); printf("┘\n");
-    printf("\n");
-
-    char* history[REPL_MAX_LINES];
-    int history_count = 0;
-    char input[16384] = {0};
-    char line[REPL_LINE_LEN];
-    int brace_depth = 0;
-
-    char ae_file[1024], c_file[1024], exe_file[1024];
-    snprintf(ae_file,  sizeof(ae_file),  "%s/_aether_repl_%d.ae",  get_temp_dir(), (int)getpid());
-    snprintf(c_file,   sizeof(c_file),   "%s/_aether_repl_%d.c",   get_temp_dir(), (int)getpid());
-    snprintf(exe_file, sizeof(exe_file), "%s/_aether_repl_%d" EXE_EXT, get_temp_dir(), (int)getpid());
-
-    while (1) {
-        if (brace_depth > 0)
-            printf("...  ");
-        else if (input[0])
-            printf("  .. ");
-        else
-            printf("  ae> ");
-        fflush(stdout);
-        if (!fgets(line, sizeof(line), stdin)) break;
-        line[strcspn(line, "\r\n")] = '\0';
-
-        // Commands (only at top level, not mid-block)
-        if (brace_depth == 0 && !input[0]) {
-            if (strcmp(line, ":quit") == 0 || strcmp(line, ":q") == 0 ||
-                strcmp(line, "exit") == 0  || strcmp(line, "quit") == 0) break;
-            if (strcmp(line, ":help") == 0 || strcmp(line, ":h") == 0) {
-                printf("\n");
-                printf("  Commands:\n");
-                printf("    :help  :h    Show this help\n");
-                printf("    :reset :r    Clear session state\n");
-                printf("    :show  :s    Show session code\n");
-                printf("    :quit  :q    Exit (also: quit, exit)\n");
-                printf("\n");
-                printf("  Usage:\n");
-                printf("    Single lines run immediately:\n");
-                printf("      ae> println(\"hello\")\n");
-                printf("      hello\n");
-                printf("\n");
-                printf("    Assignments persist across evaluations:\n");
-                printf("      ae> x = 5\n");
-                printf("      ae> println(x + 1)\n");
-                printf("      6\n");
-                printf("\n");
-                printf("    Multi-line blocks auto-continue until braces close:\n");
-                printf("      ae> if x > 3 {\n");
-                printf("      ...   println(\"big\")\n");
-                printf("      ... }\n");
-                printf("      big\n");
-                printf("\n");
-                continue;
-            }
-            if (strcmp(line, ":reset") == 0 || strcmp(line, ":r") == 0) {
-                for (int i = 0; i < history_count; i++) free(history[i]);
-                history_count = 0;
-                printf("  Session reset.\n");
-                continue;
-            }
-            if (strcmp(line, ":show") == 0 || strcmp(line, ":s") == 0) {
-                if (history_count == 0) { printf("  (empty session)\n"); continue; }
-                printf("\n");
-                for (int i = 0; i < history_count; i++)
-                    printf("    %s\n", history[i]);
-                printf("\n");
-                continue;
-            }
-        }
-
-        // Track brace depth
-        int prev_depth = brace_depth;
-        for (char* p = line; *p; p++) {
-            if (*p == '{') brace_depth++;
-            else if (*p == '}' && brace_depth > 0) brace_depth--;
-        }
-        int is_empty = (strlen(line) == 0);
-        int block_closed = (prev_depth > 0 && brace_depth == 0);
-
-        // Accumulate non-empty lines
-        if (!is_empty) {
-            if (input[0]) strncat(input, "\n", sizeof(input) - strlen(input) - 1);
-            strncat(input, line, sizeof(input) - strlen(input) - 1);
-        }
-
-        // Decide when to execute:
-        // 1. Block just closed (multi-line if/while/for)
-        // 2. Empty line with pending input (explicit trigger)
-        // 3. Single complete line (no open braces, no prior accumulation)
-        int should_run = 0;
-        if (block_closed && input[0])
-            should_run = 1;
-        else if (is_empty && input[0])
-            should_run = 1;
-        else if (!is_empty && brace_depth == 0 && prev_depth == 0 &&
-                 !strchr(input, '\n') && repl_is_complete_line(input))
-            should_run = 1;
-
-        if (should_run) {
-            if (repl_eval(ae_file, c_file, exe_file, history,
-                          history_count, input)) {
-                repl_persist(history, &history_count, input);
-            }
-            input[0] = '\0';
-            brace_depth = 0;
-        }
-    }
-
-    for (int i = 0; i < history_count; i++) free(history[i]);
-    remove(ae_file);
-    remove(c_file);
-    remove(exe_file);
-    printf("\n  Goodbye!\n\n");
-    return 0;
-}
-
-// --------------------------------------------------------------------------
-// Version manager: list available releases, install, and switch versions
-// --------------------------------------------------------------------------
-
-// Compile-time platform string used to pick the right release archive.
-#if defined(_WIN32)
-#  if defined(__aarch64__) || defined(_M_ARM64)
-#    define AE_PLATFORM "windows-arm64"
-#  else
-#    define AE_PLATFORM "windows-x86_64"
-#  endif
-#  define AE_ARCHIVE_EXT ".zip"
-#elif defined(__APPLE__) && (defined(__arm64__) || defined(__aarch64__))
-#  define AE_PLATFORM "macos-arm64"
-#  define AE_ARCHIVE_EXT ".tar.gz"
-#elif defined(__APPLE__)
-#  define AE_PLATFORM "macos-x86_64"
-#  define AE_ARCHIVE_EXT ".tar.gz"
-#elif defined(__linux__) && (defined(__aarch64__) || defined(__arm64__))
-#  define AE_PLATFORM "linux-arm64"
-#  define AE_ARCHIVE_EXT ".tar.gz"
-#else
-#  define AE_PLATFORM "linux-x86_64"
-#  define AE_ARCHIVE_EXT ".tar.gz"
-#endif
-
-#define AE_GITHUB_REPO "nicolasmd87/aether"
-
-// Download url → dest file. Uses curl/wget on POSIX, PowerShell on Windows.
-// Creates parent directories of dest if they don't exist.
-static int ae_download(const char* url, const char* dest) {
-    // Ensure parent directory exists (e.g. ~/.aether/ for releases.json)
-    {
-        char parent[1024];
-        strncpy(parent, dest, sizeof(parent) - 1);
-        parent[sizeof(parent) - 1] = '\0';
-        char* slash = strrchr(parent, '/');
-        if (!slash) slash = strrchr(parent, '\\');
-        if (slash) { *slash = '\0'; mkdirs(parent); }
-    }
-#ifdef _WIN32
-    char tmp[64];
-    snprintf(tmp, sizeof(tmp), "ae_dl_%u.ps1", (unsigned)GetCurrentProcessId());
-    char ps_path[1024];
-    snprintf(ps_path, sizeof(ps_path), "%s\\%s", get_temp_dir(), tmp);
-    FILE* ps = fopen(ps_path, "w");
-    if (!ps) return 1;
-    fprintf(ps,
-        "$ProgressPreference='SilentlyContinue'\n"
-        "Invoke-WebRequest -Uri '%s' -OutFile '%s' "
-        "-Headers @{'User-Agent'='ae-cli'}\n",
-        url, dest);
-    fclose(ps);
-    char cmd[2048];
-    snprintf(cmd, sizeof(cmd),
-        "powershell -NoProfile -ExecutionPolicy Bypass -File \"%s\" >nul 2>&1", ps_path);
-    int r = system(cmd);
-    remove(ps_path);
-    return r;
-#else
-    char cmd[2048];
-    if (system("curl --version >/dev/null 2>&1") == 0)
-        snprintf(cmd, sizeof(cmd), "curl -fsSL -o \"%s\" \"%s\" 2>/dev/null", dest, url);
-    else
-        snprintf(cmd, sizeof(cmd), "wget -q --no-verbose -O \"%s\" \"%s\" 2>/dev/null", dest, url);
-    return system(cmd);
-#endif
-}
-
-// Extract archive → dest_dir.
-static int ae_extract(const char* archive, const char* dest_dir) {
-#ifdef _WIN32
-    char tmp[64];
-    snprintf(tmp, sizeof(tmp), "ae_ex_%u.ps1", (unsigned)GetCurrentProcessId());
-    char ps_path[1024];
-    snprintf(ps_path, sizeof(ps_path), "%s\\%s", get_temp_dir(), tmp);
-    FILE* ps = fopen(ps_path, "w");
-    if (!ps) return 1;
-    fprintf(ps,
-        "$ProgressPreference='SilentlyContinue'\n"
-        "Expand-Archive -Path '%s' -DestinationPath '%s' -Force\n",
-        archive, dest_dir);
-    fclose(ps);
-    char cmd[2048];
-    snprintf(cmd, sizeof(cmd),
-        "powershell -NoProfile -ExecutionPolicy Bypass -File \"%s\" >nul 2>&1", ps_path);
-    int r = system(cmd);
-    remove(ps_path);
-    return r;
-#else
-    char cmd[2048];
-    snprintf(cmd, sizeof(cmd), "tar -xzf \"%s\" -C \"%s\"", archive, dest_dir);
-    return system(cmd);
-#endif
-}
-
-// Clear macOS Gatekeeper state on a freshly installed/copied binary.
-//
-// Released Aether binaries are adhoc-signed and get a quarantine xattr
-// the moment they're downloaded. Clearing the quarantine alone is not
-// enough: Gatekeeper also caches an "assessment" per file, and an
-// ad-hoc-signed binary from an untrusted source is rejected by default.
-// The rejection manifests as the shell seeing "Killed: 9" (SIGKILL) or
-// `ae` hanging for seconds while `syspolicyd` evaluates the binary.
-//
-// The reliable fix is to re-sign in place with `codesign --force --sign -`.
-// This refreshes the CDHash and creates a local ad-hoc signature that
-// Gatekeeper allows to execute without notarization checks. Combined with
-// clearing the quarantine attribute, the binary runs cleanly right away.
-//
-// No-op on Linux and Windows, which have no Gatekeeper equivalent.
-static void macos_prepare_binary(const char* path) {
-#ifdef __APPLE__
-    if (!path || !*path) return;
-    char cmd[2048];
-    // Re-sign in place. Suppress both success and failure output so a
-    // missing codesign (unusual but not impossible on stripped systems)
-    // degrades to quarantine-clear only.
-    snprintf(cmd, sizeof(cmd),
-             "codesign --force --sign - \"%s\" >/dev/null 2>&1", path);
-    (void)system(cmd);
-    // Clear quarantine + any other resource forks/xattrs that would
-    // otherwise trigger an extra syspolicyd round-trip on first run.
-    snprintf(cmd, sizeof(cmd), "xattr -cr \"%s\" >/dev/null 2>&1", path);
-    (void)system(cmd);
-#else
-    (void)path;
-#endif
-}
-
-// Prepare every executable in a bin directory. Used after install/extract
-// and after copying binaries into ~/.aether/bin/ from a versioned install.
-static void macos_prepare_bin_dir(const char* bin_dir) {
-#ifdef __APPLE__
-    if (!bin_dir || !*bin_dir) return;
-    char cmd[2048];
-    // find + xargs handles spaces via -print0/-0. We only touch regular
-    // files; symlinks are skipped so we don't re-sign a link's target.
-    snprintf(cmd, sizeof(cmd),
-             "find \"%s\" -maxdepth 1 -type f -perm +111 -print0 2>/dev/null "
-             "| xargs -0 -I {} codesign --force --sign - \"{}\" >/dev/null 2>&1",
-             bin_dir);
-    (void)system(cmd);
-    snprintf(cmd, sizeof(cmd), "xattr -cr \"%s\" >/dev/null 2>&1", bin_dir);
-    (void)system(cmd);
-#else
-    (void)bin_dir;
-#endif
-}
-
-// List available releases from GitHub. Marks installed + current versions.
-static int cmd_version_list(void) {
-    const char* home = get_home_dir();
-
-    // Determine which version is actually active.
-    // Priority: 1) ~/.aether/active_version file (authoritative — written by install.sh and ae version use)
-    //           2) ~/.aether/current symlink (legacy fallback)
-    //           3) compiled-in AE_VERSION (fallback)
-    char active_ver[64] = "";
-
-    // Check active_version file first (always authoritative)
-    {
-        char avpath[512];
-#ifdef _WIN32
-        snprintf(avpath, sizeof(avpath), "%s\\.aether\\active_version", home);
-#else
-        snprintf(avpath, sizeof(avpath), "%s/.aether/active_version", home);
-#endif
-        FILE* avf = fopen(avpath, "r");
-        if (avf) {
-            if (fgets(active_ver, sizeof(active_ver), avf)) {
-                char* nl = strchr(active_ver, '\n'); if (nl) *nl = '\0';
-                char* cr = strchr(active_ver, '\r'); if (cr) *cr = '\0';
-            }
-            fclose(avf);
-        }
-    }
-
-#ifndef _WIN32
-    // Legacy fallback: resolve ~/.aether/current symlink
-    if (active_ver[0] == '\0') {
-        char current_link[512], target[1024];
-        snprintf(current_link, sizeof(current_link), "%s/.aether/current", home);
-        ssize_t rlen = readlink(current_link, target, sizeof(target) - 1);
-        if (rlen > 0) {
-            target[rlen] = '\0';
-            const char* last = strrchr(target, '/');
-            if (last) last++; else last = target;
-            if (last[0] == 'v') last++;
-            strncpy(active_ver, last, sizeof(active_ver) - 1);
-            active_ver[sizeof(active_ver) - 1] = '\0';
-        }
-    }
-#endif
-
-    // Fallback: use compiled-in version
-    if (active_ver[0] == '\0') {
-        strncpy(active_ver, AE_VERSION, sizeof(active_ver) - 1);
-        active_ver[sizeof(active_ver) - 1] = '\0';
-    }
-
-    // Fetch the GitHub releases JSON into a temp file
-    char json_path[512];
-#ifdef _WIN32
-    snprintf(json_path, sizeof(json_path), "%s\\.aether\\releases.json", home);
-#else
-    snprintf(json_path, sizeof(json_path), "%s/ae_releases_%d.json", get_temp_dir(), (int)getpid());
-#endif
-    char url[256];
-    snprintf(url, sizeof(url),
-        "https://api.github.com/repos/" AE_GITHUB_REPO "/releases?per_page=20");
-
-    printf("Fetching release list...\n");
-    if (ae_download(url, json_path) != 0) {
-        fprintf(stderr, "Failed to fetch releases. Check your internet connection.\n");
-        return 1;
-    }
-
-    FILE* f = fopen(json_path, "r");
-    if (!f) {
-        fprintf(stderr, "Failed to read release data.\n");
-        return 1;
-    }
-
-    printf("\nAvailable Aether releases  (platform: " AE_PLATFORM "):\n\n");
-    printf("  %-16s  %s\n", "Version", "Status");
-    printf("  %-16s  %s\n", "-------", "------");
-
-    // Read whole file, scan for "tag_name" occurrences
-    char buf[131072];
-    size_t n = fread(buf, 1, sizeof(buf) - 1, f);
-    fclose(f);
-    remove(json_path);
-    buf[n] = '\0';
-
-    int found = 0;
-    char* p = buf;
-    while ((p = strstr(p, "\"tag_name\"")) != NULL) {
-        p += 10;  // skip past "tag_name"
-        char* q = strchr(p, '"'); if (!q) break; q++;   // opening "
-        char* end = strchr(q, '"'); if (!end) break;
-        size_t len = (size_t)(end - q);
-        if (len == 0 || len > 32) { p = end + 1; continue; }
-
-        char tag[33];
-        memcpy(tag, q, len);
-        tag[len] = '\0';
-        p = end + 1;
-
-        // v-prefix normalisation: strip 'v' to compare with active version
-        const char* ver = (tag[0] == 'v') ? tag + 1 : tag;
-        bool is_current = strcmp(ver, active_ver) == 0;
-
-        // Check locally installed
-        char ver_dir[512];
-        snprintf(ver_dir, sizeof(ver_dir), "%s/.aether/versions/%s", home, tag);
-        bool installed = dir_exists(ver_dir);
-
-        const char* status = is_current ? "* current"
-                           : installed  ? "  installed"
-                                        : "";
-        printf("  %-16s  %s\n", tag, status);
-        found++;
-    }
-
-    if (!found) {
-        printf("  (no releases found)\n");
-    }
-    printf("\n");
-    // Show latest found tag in examples (or fallback)
-    if (found > 0) {
-        // First tag found is the latest (GitHub returns newest first)
-        // Re-scan to get it
-        char latest[33] = "v0.1.0";
-        char* lp = strstr(buf, "\"tag_name\"");
-        if (lp) {
-            lp += 10;
-            char* lq = strchr(lp, '"'); if (lq) { lq++;
-            char* le = strchr(lq, '"'); if (le) {
-                size_t ll = (size_t)(le - lq);
-                if (ll > 0 && ll < sizeof(latest)) { memcpy(latest, lq, ll); latest[ll] = '\0'; }
-            }}
-        }
-        printf("Install a version:  ae version install %s\n", latest);
-        printf("Switch versions:    ae version use %s\n", latest);
-    } else {
-        printf("Install a version:  ae version install <version>\n");
-        printf("Switch versions:    ae version use <version>\n");
-    }
-    return 0;
-}
-
-// Download and install a specific version into ~/.aether/versions/<tag>/
-static int cmd_version_install(const char* version) {
-    char vtag[64];
-    if (version[0] != 'v') snprintf(vtag, sizeof(vtag), "v%s", version);
-    else { strncpy(vtag, version, sizeof(vtag) - 1); vtag[sizeof(vtag)-1] = '\0'; }
-
-    const char* ver = vtag + 1;  // strip leading 'v'
-    const char* home = get_home_dir();
-
-    char ver_dir[512];
-    snprintf(ver_dir, sizeof(ver_dir), "%s/.aether/versions/%s", home, vtag);
-    if (dir_exists(ver_dir)) {
-        // Verify the install is complete by checking for binaries
-        char probe[1024];
-        int has_binary = 0;
-#ifdef _WIN32
-        snprintf(probe, sizeof(probe), "%s\\bin\\aetherc.exe", ver_dir);
-        if (path_exists(probe)) has_binary = 1;
-        snprintf(probe, sizeof(probe), "%s\\aetherc.exe", ver_dir);
-        if (path_exists(probe)) has_binary = 1;
-#else
-        snprintf(probe, sizeof(probe), "%s/bin/aetherc", ver_dir);
-        if (path_exists(probe)) has_binary = 1;
-        snprintf(probe, sizeof(probe), "%s/aetherc", ver_dir);
-        if (path_exists(probe)) has_binary = 1;
-#endif
-        if (has_binary) {
-            // Also verify the install has sources or a prebuilt lib —
-            // old ae versions had an extraction bug that only copied bin/
-            char lib_probe[1024], share_probe[1024];
-            int has_sources = 0;
-            snprintf(lib_probe, sizeof(lib_probe), "%s/lib/aether/libaether.a", ver_dir);
-            if (path_exists(lib_probe)) has_sources = 1;
-            snprintf(share_probe, sizeof(share_probe), "%s/share/aether/runtime", ver_dir);
-            if (dir_exists(share_probe)) has_sources = 1;
-            if (has_sources) {
-                printf("Version %s is already installed.\n", vtag);
-                printf("Switch to it with: ae version use %s\n", vtag);
-                return 0;
-            }
-            printf("Version %s has binaries but missing lib/share — reinstalling...\n", vtag);
-            // Fall through to remove and re-download
-        }
-        // Incomplete install — remove and re-download
-        printf("Incomplete installation of %s detected, reinstalling...\n", vtag);
-#ifdef _WIN32
-        char rm_cmd[1024];
-        snprintf(rm_cmd, sizeof(rm_cmd), "rmdir /S /Q \"%s\"", ver_dir);
-        if (system(rm_cmd) != 0) {
-            fprintf(stderr, "Warning: failed to remove incomplete install at %s\n", ver_dir);
-        }
-#else
-        char rm_cmd[1024];
-        snprintf(rm_cmd, sizeof(rm_cmd), "rm -rf '%s'", ver_dir);
-        if (system(rm_cmd) != 0) {
-            fprintf(stderr, "Warning: failed to remove incomplete install at %s\n", ver_dir);
-        }
-#endif
-    }
-
-    // Build URL and local archive path
-    char filename[256], url[1024], archive[512];
-    snprintf(filename, sizeof(filename),
-        "aether-%s-" AE_PLATFORM AE_ARCHIVE_EXT, ver);
-    snprintf(url, sizeof(url),
-        "https://github.com/" AE_GITHUB_REPO "/releases/download/%s/%s",
-        vtag, filename);
-    snprintf(archive, sizeof(archive), "%s/.aether/%s", home, filename);
-
-    printf("Downloading Aether %s for " AE_PLATFORM "...\n", vtag);
-    fflush(stdout);
-    if (ae_download(url, archive) != 0) {
-        fprintf(stderr, "Error: Version %s not found for " AE_PLATFORM ".\n", vtag);
-        fprintf(stderr, "Run 'ae version list' to see available versions.\n");
-        return 1;
-    }
-
-    // Verify the downloaded file is a real archive, not a 404 HTML page.
-    // Valid archives are at least 10KB; GitHub 404 pages are ~10-20KB HTML
-    // but tar.gz/zip archives for Aether are always >100KB.
-    {
-        FILE* af = fopen(archive, "rb");
-        if (!af) {
-            fprintf(stderr, "Error: Downloaded file not found.\n");
-            return 1;
-        }
-        fseek(af, 0, SEEK_END);
-        long asize = ftell(af);
-        // Also check the first bytes for archive magic
-        fseek(af, 0, SEEK_SET);
-        unsigned char magic[4] = {0};
-        if (fread(magic, 1, 4, af) < 4) { /* short read — magic stays zeroed */ }
-        fclose(af);
-
-        int is_gzip = (magic[0] == 0x1f && magic[1] == 0x8b);  // .tar.gz
-        int is_zip  = (magic[0] == 'P' && magic[1] == 'K');     // .zip
-        int is_xz   = (magic[0] == 0xFD && magic[1] == '7');    // .tar.xz
-
-        if (!is_gzip && !is_zip && !is_xz) {
-            remove(archive);
-            fprintf(stderr, "Error: Version %s not found for platform " AE_PLATFORM ".\n", vtag);
-            fprintf(stderr, "The download returned an error page, not a release archive.\n");
-            fprintf(stderr, "Available versions: ae version list\n");
-            return 1;
-        }
-        if (asize < 1024) {
-            remove(archive);
-            fprintf(stderr, "Error: Downloaded archive is too small (%ld bytes) — likely corrupt.\n", asize);
-            return 1;
-        }
-    }
-
-    mkdirs(ver_dir);
-    printf("Extracting...\n");
-
-    // The release archive contains a top-level directory (e.g. "release/").
-    // Extract to a temp dir first, then move the contents into ver_dir.
-    char tmp_dir[512];
-    snprintf(tmp_dir, sizeof(tmp_dir), "%s/.aether/_tmp_install", home);
-    mkdirs(tmp_dir);
-
-    if (ae_extract(archive, tmp_dir) != 0) {
-        fprintf(stderr, "Extraction failed.\n");
-        remove(archive);
-        return 1;
-    }
-    remove(archive);
-
-    // Move extracted contents into ver_dir.
-    // Release archives may have a single wrapper directory (e.g. "aether-v0.21.0-macos-arm64/")
-    // OR may have bin/, lib/, share/, include/ directly at root. Handle both cases.
-#ifdef _WIN32
-    char cmd[2048];
-    snprintf(cmd, sizeof(cmd), "xcopy /E /Y /Q \"%s\\*\" \"%s\\\"", tmp_dir, ver_dir);
-    if (system(cmd) != 0) {
-        fprintf(stderr, "Error: Failed to copy installation files.\n");
-        return 1;
-    }
-    snprintf(cmd, sizeof(cmd), "rmdir /S /Q \"%s\"", tmp_dir);
-    if (system(cmd) != 0) { /* non-fatal: temp dir cleanup */ }
-#else
-    {
-        char cmd[4096];
-        // If there is exactly one top-level entry and it is a directory,
-        // treat it as a wrapper and copy its contents. Otherwise copy
-        // everything directly (the archive has bin/, lib/, etc. at root).
-        snprintf(cmd, sizeof(cmd),
-            "entries=$(ls '%s' | wc -l | tr -d ' '); "
-            "single=$(ls -d '%s'/*/ 2>/dev/null | wc -l | tr -d ' '); "
-            "if [ \"$entries\" = \"1\" ] && [ \"$single\" = \"1\" ]; then "
-            "  src=$(ls -d '%s'/*/); cp -r \"$src\"* '%s/'; "
-            "else "
-            "  cp -r '%s'/* '%s/'; "
-            "fi",
-            tmp_dir, tmp_dir, tmp_dir, ver_dir, tmp_dir, ver_dir);
-        if (system(cmd) != 0) {
-            fprintf(stderr, "Error: Failed to copy installation files.\n");
-            return 1;
-        }
-        snprintf(cmd, sizeof(cmd), "rm -rf '%s'", tmp_dir);
-        if (system(cmd) != 0) { /* non-fatal: temp dir cleanup */ }
-    }
-#endif
-
-    // Verify the installation has the expected structure.
-    // Releases should have bin/, lib/ or share/aether/ — if we only see
-    // flat binaries, the extraction went wrong.
-    {
-        char probe[1024];
-        int has_structure = 0;
-        snprintf(probe, sizeof(probe), "%s/bin", ver_dir);
-        if (dir_exists(probe)) has_structure = 1;
-        snprintf(probe, sizeof(probe), "%s/lib", ver_dir);
-        if (dir_exists(probe)) has_structure = 1;
-        snprintf(probe, sizeof(probe), "%s/share/aether", ver_dir);
-        if (dir_exists(probe)) has_structure = 1;
-        if (!has_structure) {
-            // Clean up the empty/broken install directory
-#ifdef _WIN32
-            snprintf(probe, sizeof(probe), "rmdir /S /Q \"%s\"", ver_dir);
-#else
-            snprintf(probe, sizeof(probe), "rm -rf '%s'", ver_dir);
-#endif
-            if (system(probe) != 0) { /* cleanup failed — non-fatal */ }
-            fprintf(stderr, "Error: Installation of %s failed — no bin/, lib/, or share/ found.\n", vtag);
-            fprintf(stderr, "This version may not have a release for " AE_PLATFORM ".\n");
-            fprintf(stderr, "Available versions: ae version list\n");
-            return 1;
-        }
-    }
-
-    // #959: integrity-check the prebuilt runtime archive after extraction.
-    // An interrupted/partial extract has been observed to leave a truncated
-    // libaether.a whose symbols are undefined (U) instead of defined (T), so
-    // every later `ae build` fails to link with nothing pointing at the cause.
-    // Validate that the archive (canonical nested path, or the flat fallback
-    // some packages ship) is a well-formed `ar` archive of plausible size; on
-    // failure remove the install and tell the user to retry, rather than
-    // leaving a landmine. Skipped when the package ships sources only (no
-    // prebuilt archive present). This does not replace a full checksum verify
-    // (which needs the release pipeline to publish per-asset sums) but catches
-    // the truncation that actually corrupted a macOS download.
-    {
-        char arch_path[1024];
-        snprintf(arch_path, sizeof(arch_path), "%s/lib/aether/libaether.a", ver_dir);
-        if (!path_exists(arch_path))
-            snprintf(arch_path, sizeof(arch_path), "%s/lib/libaether.a", ver_dir);
-        if (path_exists(arch_path)) {
-            FILE* lf = fopen(arch_path, "rb");
-            int ok = 0;
-            if (lf) {
-                char hdr[8] = {0};
-                size_t n = fread(hdr, 1, sizeof(hdr), lf);
-                fseek(lf, 0, SEEK_END);
-                long lsize = ftell(lf);
-                fclose(lf);
-                /* `ar` archives begin with the 8-byte magic "!<arch>\n" on
-                 * every platform we ship (GNU/BSD/macOS ar). A complete
-                 * runtime+stdlib archive is multi-megabyte; a truncated one
-                 * falls well short of this conservative 64 KB floor. */
-                ok = (n == sizeof(hdr) &&
-                      memcmp(hdr, "!<arch>\n", sizeof(hdr)) == 0 &&
-                      lsize > 65536);
-            }
-            if (!ok) {
-                char rmc[1024];
-#ifdef _WIN32
-                snprintf(rmc, sizeof(rmc), "rmdir /S /Q \"%s\"", ver_dir);
-#else
-                snprintf(rmc, sizeof(rmc), "rm -rf '%s'", ver_dir);
-#endif
-                if (system(rmc) != 0) { /* cleanup failed — non-fatal */ }
-                fprintf(stderr,
-                    "Error: %s installed a corrupt runtime archive "
-                    "(lib/.../libaether.a is truncated or not an `ar` archive).\n",
-                    vtag);
-                fprintf(stderr,
-                    "The download or extraction was likely interrupted. "
-                    "Re-run: ae version install %s\n", vtag);
-                return 1;
-            }
-        }
-    }
-
-    // Prepare binaries for macOS Gatekeeper. Download + extract leaves the
-    // binaries quarantined; without this step the first run of any binary
-    // from the fresh install (including the self-invocation from
-    // `ae version use`) hangs or gets SIGKILL'd by syspolicyd.
-    {
-        char bin_sub[1024];
-        snprintf(bin_sub, sizeof(bin_sub), "%s/bin", ver_dir);
-        if (dir_exists(bin_sub)) {
-            macos_prepare_bin_dir(bin_sub);
-        }
-    }
-
-    printf("Installed Aether %s → %s\n", vtag, ver_dir);
-    printf("Switch to it with: ae version use %s\n", vtag);
-    return 0;
-}
-
-// Determine where binaries live inside a version directory.
-// Release archives may have a bin/ subdirectory or binaries at root.
-static void resolve_version_bin_dir(const char* ver_dir, char* out, size_t outsz) {
-    char probe[1024];
-#ifdef _WIN32
-    snprintf(probe, sizeof(probe), "%s\\bin\\aetherc" EXE_EXT, ver_dir);
-    if (path_exists(probe)) { snprintf(out, outsz, "%s\\bin", ver_dir); return; }
-    snprintf(probe, sizeof(probe), "%s\\bin\\ae" EXE_EXT, ver_dir);
-    if (path_exists(probe)) { snprintf(out, outsz, "%s\\bin", ver_dir); return; }
-    snprintf(out, outsz, "%s", ver_dir);
-#else
-    snprintf(probe, sizeof(probe), "%s/bin/aetherc" EXE_EXT, ver_dir);
-    if (path_exists(probe)) { snprintf(out, outsz, "%s/bin", ver_dir); return; }
-    snprintf(probe, sizeof(probe), "%s/bin/ae" EXE_EXT, ver_dir);
-    if (path_exists(probe)) { snprintf(out, outsz, "%s/bin", ver_dir); return; }
-    snprintf(out, outsz, "%s", ver_dir);
-#endif
-}
-
-static int cmd_version_use(const char* version) {
-    char vtag[64];
-    if (version[0] != 'v') snprintf(vtag, sizeof(vtag), "v%s", version);
-    else { strncpy(vtag, version, sizeof(vtag) - 1); vtag[sizeof(vtag)-1] = '\0'; }
-
-    const char* home = get_home_dir();
-    char ver_dir[512];
-    snprintf(ver_dir, sizeof(ver_dir), "%s/.aether/versions/%s", home, vtag);
-
-    if (!dir_exists(ver_dir)) {
-        fprintf(stderr, "Version %s is not installed.\n", vtag);
-        fprintf(stderr, "Install it first: ae version install %s\n", vtag);
-        return 1;
-    }
-
-    char src_bin[1024];
-    resolve_version_bin_dir(ver_dir, src_bin, sizeof(src_bin));
-
-#ifdef _WIN32
-    // Backup the currently active version to versions/ before overwriting.
-    // This preserves the initial install (e.g., v0.30.0 installed via install.sh
-    // lives in ~/.aether/ directly, not in versions/).
-    {
-        char avpath_bak[512];
-        snprintf(avpath_bak, sizeof(avpath_bak), "%s\\.aether\\active_version", home);
-        FILE* avf_bak = fopen(avpath_bak, "r");
-        if (avf_bak) {
-            char cur_ver[64] = "";
-            if (fgets(cur_ver, sizeof(cur_ver), avf_bak)) {
-                char* nl = strchr(cur_ver, '\n');
-                if (nl) *nl = '\0';
-            }
-            fclose(avf_bak);
-            if (cur_ver[0]) {
-                char cur_vtag[66];
-                if (cur_ver[0] != 'v') snprintf(cur_vtag, sizeof(cur_vtag), "v%s", cur_ver);
-                else { strncpy(cur_vtag, cur_ver, sizeof(cur_vtag) - 1); cur_vtag[sizeof(cur_vtag)-1] = '\0'; }
-                char cur_ver_dir[512];
-                snprintf(cur_ver_dir, sizeof(cur_ver_dir), "%s\\.aether\\versions\\%s", home, cur_vtag);
-                if (!dir_exists(cur_ver_dir)) {
-                    // Current version not in versions/ — back it up
-                    char bak_cmd[2048];
-                    char dest_root_bak[512];
-                    snprintf(dest_root_bak, sizeof(dest_root_bak), "%s\\.aether", home);
-                    mkdirs(cur_ver_dir);
-                    snprintf(bak_cmd, sizeof(bak_cmd),
-                        "robocopy \"%s\" \"%s\" /E /NFL /NDL /NJH /NJS /IS /IT /XD versions cache >nul 2>&1",
-                        dest_root_bak, cur_ver_dir);
-                    if (system(bak_cmd) != 0) { /* backup failed — non-fatal */ }
-                }
-            }
-        }
-    }
-
-    // Copy the entire version directory to ~/.aether/ so lib/, include/,
-    // share/ are available alongside bin/.
-    char dest_root[512];
-    snprintf(dest_root, sizeof(dest_root), "%s\\.aether", home);
-    char cmd[2048];
-    // robocopy /E copies all subdirectories; /NFL /NDL /NJH /NJS suppress output
-    snprintf(cmd, sizeof(cmd),
-        "robocopy \"%s\" \"%s\" /E /NFL /NDL /NJH /NJS /IS /IT >nul 2>&1",
-        ver_dir, dest_root);
-    int rc = system(cmd);
-    // robocopy returns 0-7 for success, >=8 for failure
-    if (rc >= 8) {
-        // Fall back to xcopy
-        snprintf(cmd, sizeof(cmd),
-            "xcopy /E /Y /Q \"%s\\*\" \"%s\\\"", ver_dir, dest_root);
-        if (system(cmd) != 0) {
-            fprintf(stderr, "Failed to copy version files from %s to %s\n", ver_dir, dest_root);
-            return 1;
-        }
-    }
-#else
-    // Backup current version to versions/ before switching (preserves initial install)
-    {
-        char avpath_bak[512];
-        snprintf(avpath_bak, sizeof(avpath_bak), "%s/.aether/active_version", home);
-        FILE* avf_bak = fopen(avpath_bak, "r");
-        if (avf_bak) {
-            char cur_ver[64] = "";
-            if (fgets(cur_ver, sizeof(cur_ver), avf_bak)) {
-                char* nl = strchr(cur_ver, '\n'); if (nl) *nl = '\0';
-            }
-            fclose(avf_bak);
-            if (cur_ver[0]) {
-                char cur_vtag[64];
-                if (cur_ver[0] != 'v') snprintf(cur_vtag, sizeof(cur_vtag), "v%s", cur_ver);
-                else { strncpy(cur_vtag, cur_ver, sizeof(cur_vtag) - 1); cur_vtag[sizeof(cur_vtag)-1] = '\0'; }
-                char cur_ver_dir[512];
-                snprintf(cur_ver_dir, sizeof(cur_ver_dir), "%s/.aether/versions/%s", home, cur_vtag);
-                if (!dir_exists(cur_ver_dir)) {
-                    char bak_cmd[4096];
-                    mkdirs(cur_ver_dir);
-                    // Copy bin/, lib/, include/, share/ but NOT versions/ or cache/
-                    snprintf(bak_cmd, sizeof(bak_cmd),
-                        "cp -r \"%s/.aether/bin\" \"%s/\" 2>/dev/null; "
-                        "cp -r \"%s/.aether/lib\" \"%s/\" 2>/dev/null; "
-                        "cp -r \"%s/.aether/include\" \"%s/\" 2>/dev/null; "
-                        "cp -r \"%s/.aether/share\" \"%s/\" 2>/dev/null; true",
-                        home, cur_ver_dir, home, cur_ver_dir,
-                        home, cur_ver_dir, home, cur_ver_dir);
-                    if (system(bak_cmd) != 0) { /* backup failed — non-fatal */ }
-                }
-            }
-        }
-    }
-
-    // Verify the source bin dir actually contains an `ae` binary before
-    // we start mutating anything. This catches extraction-layout bugs
-    // early rather than leaving the user with a half-switched install.
-    {
-        // src_bin is 1024 bytes; adding "/ae" + EXE_EXT (".exe" worst
-        // case) + NUL needs 8 bytes of headroom over the source.
-        char src_ae[1040];
-        snprintf(src_ae, sizeof(src_ae), "%s/ae" EXE_EXT, src_bin);
-        if (!path_exists(src_ae)) {
-            fprintf(stderr, "Error: no ae binary at %s. The install for %s looks incomplete.\n", src_ae, vtag);
-            fprintf(stderr, "Try reinstalling: ae version install %s\n", vtag);
-            return 1;
-        }
-    }
-
-    // POSIX: update ~/.aether/current symlink
-    char current[512];
-    snprintf(current, sizeof(current), "%s/.aether/current", home);
-    remove(current);
-    char cmd[4096];
-    snprintf(cmd, sizeof(cmd), "ln -sf \"%s\" \"%s\"", ver_dir, current);
-    if (system(cmd) != 0) {
-        fprintf(stderr, "Failed to create symlink. Try manually:\n");
-        fprintf(stderr, "  ln -sf %s %s\n", ver_dir, current);
-        return 1;
-    }
-
-    // Copy binaries into ~/.aether/bin/. Previously this was a single
-    // `cp -f "%s"/* "%s/" 2>/dev/null; true` which suppressed every
-    // possible failure mode including a missing source or permissions
-    // error. We now fail loudly and verify afterwards.
-    char dest_bin[512];
-    snprintf(dest_bin, sizeof(dest_bin), "%s/.aether/bin", home);
-    snprintf(cmd, sizeof(cmd), "mkdir -p \"%s\"", dest_bin);
-    if (system(cmd) != 0) {
-        fprintf(stderr, "Error: failed to create %s\n", dest_bin);
-        return 1;
-    }
-    // Use a subshell with nullglob so an empty source (which shouldn't
-    // happen after the verification above, but just in case) fails the
-    // `cp` explicitly rather than trying to copy a literal "*".
-    snprintf(cmd, sizeof(cmd),
-        "/bin/sh -c 'set -e; for f in \"%s\"/*; do cp -f \"$f\" \"%s/\"; done'",
-        src_bin, dest_bin);
-    if (system(cmd) != 0) {
-        fprintf(stderr, "Error: failed to copy binaries from %s to %s\n", src_bin, dest_bin);
-        return 1;
-    }
-
-    // Verify the copy landed.
-    {
-        char dest_ae[1024];
-        snprintf(dest_ae, sizeof(dest_ae), "%s/ae" EXE_EXT, dest_bin);
-        if (!path_exists(dest_ae)) {
-            fprintf(stderr, "Error: expected %s after copy, but it's not there.\n", dest_ae);
-            return 1;
-        }
-    }
-
-    // Prepare the binaries for macOS Gatekeeper (codesign + xattr clear).
-    // Without this the next invocation of `ae` hangs or gets SIGKILL'd by
-    // syspolicyd on first run. No-op on Linux.
-    macos_prepare_bin_dir(dest_bin);
-
-    // Sync lib/, include/, and share/ from the version directory to ~/.aether/
-    // so that stale files left by a previous install.sh don't shadow the
-    // version-managed files. The 'current' symlink alone is not enough because
-    // toolchain discovery may resolve the parent directory first.
-    //
-    // We used to self-invoke the freshly copied ae with --sync-from as a
-    // bootstrap for old-binary upgrades. On macOS that invocation hangs or
-    // gets SIGKILL'd by Gatekeeper before syspolicyd finishes evaluating the
-    // just-copied binary, leaving the user waiting minutes. The in-process
-    // sync below does the same work, so we drop the self-invocation entirely.
-    {
-        char dest[512];
-        const char* subdirs[] = {"lib", "include", "share"};
-        for (int i = 0; i < 3; i++) {
-            char src_sub[1024];
-            snprintf(src_sub, sizeof(src_sub), "%s/%s", ver_dir, subdirs[i]);
-            if (dir_exists(src_sub)) {
-                snprintf(dest, sizeof(dest), "%s/.aether/%s", home, subdirs[i]);
-                snprintf(cmd, sizeof(cmd),
-                    "rm -rf \"%s\" && cp -r \"%s\" \"%s\"",
-                    dest, src_sub, dest);
-                if (system(cmd) != 0) {
-                    fprintf(stderr, "Warning: failed to sync %s to %s\n", src_sub, dest);
-                }
-            }
-        }
-    }
-#endif
-
-    // Update active_version file so 'ae version list' shows the correct current
-    // even if the symlink approach fails or on Windows.
-    {
-        char avpath[512];
-#ifdef _WIN32
-        snprintf(avpath, sizeof(avpath), "%s\\.aether\\active_version", home);
-#else
-        snprintf(avpath, sizeof(avpath), "%s/.aether/active_version", home);
-#endif
-        FILE* avf = fopen(avpath, "w");
-        if (avf) {
-            // Write version without 'v' prefix
-            const char* v = (vtag[0] == 'v') ? vtag + 1 : vtag;
-            fprintf(avf, "%s\n", v);
-            fclose(avf);
-        }
-    }
-
-    printf("Switched to Aether %s.\n", vtag);
-    return 0;
-}
-
-// "ae version [list|install|use]"
-// Read the active version from ~/.aether/active_version, fall back to compiled-in
-static const char* get_active_version(void) {
-    static char active[64];
-    const char* home = get_home_dir();
-    char avpath[512];
-#ifdef _WIN32
-    snprintf(avpath, sizeof(avpath), "%s\\.aether\\active_version", home);
-#else
-    snprintf(avpath, sizeof(avpath), "%s/.aether/active_version", home);
-#endif
-    FILE* f = fopen(avpath, "r");
-    if (f) {
-        if (fgets(active, sizeof(active), f)) {
-            // Trim newline
-            char* nl = strchr(active, '\n');
-            if (nl) *nl = '\0';
-            fclose(f);
-            if (active[0]) return active;
-        }
-        fclose(f);
-    }
-    return AE_VERSION;
-}
-
-// Fetch the newest release tag (e.g. "v0.231.0") from GitHub into `out`.
-// Returns 0 on success, non-zero on network/parse failure. GitHub returns
-// releases newest-first, so the first "tag_name" is the latest — same
-// scan cmd_version_list uses.
-static int fetch_latest_release_tag(char* out, size_t outlen) {
-    char json_path[512];
-    snprintf(json_path, sizeof(json_path), "%s/ae_latest_%d.json",
-             get_temp_dir(), (int)getpid());
-    char url[256];
-    snprintf(url, sizeof(url),
-        "https://api.github.com/repos/" AE_GITHUB_REPO "/releases?per_page=5");
-    if (ae_download(url, json_path) != 0) return -1;
-    FILE* f = fopen(json_path, "r");
-    if (!f) { remove(json_path); return -1; }
-    char buf[65536];
-    size_t n = fread(buf, 1, sizeof(buf) - 1, f);
-    fclose(f);
-    remove(json_path);
-    buf[n] = '\0';
-    char* p = strstr(buf, "\"tag_name\"");
-    if (!p) return -1;
-    p += 10;
-    char* q = strchr(p, '"'); if (!q) return -1; q++;   // opening quote
-    char* e = strchr(q, '"'); if (!e) return -1;         // closing quote
-    size_t len = (size_t)(e - q);
-    if (len == 0 || len >= outlen) return -1;
-    memcpy(out, q, len);
-    out[len] = '\0';
-    return 0;
-}
-
-// "ae install [<version>]" — install a release into ~/.aether/versions/.
-// With no version, resolves and installs the latest. A discoverable
-// top-level alias for "ae version install" (matches rustup/nvm muscle
-// memory). Does NOT switch the active version — that's "ae use <v>" or
-// "ae upgrade".
-static int cmd_install(int argc, char** argv) {
-    if (argc >= 1 && argv[0] && argv[0][0]) {
-        return cmd_version_install(argv[0]);
-    }
-    char latest[64];
-    printf("Resolving latest release...\n");
-    if (fetch_latest_release_tag(latest, sizeof(latest)) != 0) {
-        fprintf(stderr, "Could not determine the latest release.\n");
-        fprintf(stderr, "Check your connection, or name a version: ae install <v>\n");
-        fprintf(stderr, "('ae version list' shows what's available.)\n");
-        return 1;
-    }
-    return cmd_version_install(latest);
-}
-
-// "ae upgrade" / "ae update" — install the latest release and switch to it.
-// The one-shot "get me the newest Aether" command; a no-op (with a notice)
-// when already current.
-static int cmd_upgrade(void) {
-    char latest[64];
-    printf("Checking for the latest release...\n");
-    if (fetch_latest_release_tag(latest, sizeof(latest)) != 0) {
-        fprintf(stderr, "Could not determine the latest release. Check your connection.\n");
-        return 1;
-    }
-    const char* latest_ver = (latest[0] == 'v') ? latest + 1 : latest;
-    const char* active = get_active_version();   // no leading 'v'
-    if (active && strcmp(active, latest_ver) == 0) {
-        printf("Already on the latest version (%s).\n", latest);
-        return 0;
-    }
-    printf("Upgrading %s -> %s\n", (active && active[0]) ? active : "(unknown)", latest);
-    int rc = cmd_version_install(latest);
-    if (rc != 0) return rc;
-    return cmd_version_use(latest);
-}
-
-static int cmd_version(int argc, char** argv) {
-    if (argc == 0) {
-        printf("ae %s (Aether Language)\n", get_active_version());
-        printf("Platform: " AE_PLATFORM "\n");
-        printf("\nSubcommands:\n");
-        printf("  ae version list              List all available releases\n");
-        printf("  ae version install <v>       Download and install a release\n");
-        printf("  ae version use <v>           Switch to an installed release\n");
-        return 0;
-    }
-    const char* sub = argv[0];
-    if (strcmp(sub, "list") == 0)    return cmd_version_list();
-    if (strcmp(sub, "install") == 0) {
-        if (argc < 2) { fprintf(stderr, "Usage: ae version install <v>\n"); return 1; }
-        return cmd_version_install(argv[1]);
-    }
-    if (strcmp(sub, "use") == 0) {
-        if (argc < 2) { fprintf(stderr, "Usage: ae version use <v>\n"); return 1; }
-        return cmd_version_use(argv[1]);
-    }
-    // Internal: called by old ae binaries after copying new ae to ~/.aether/bin/
-    // Syncs lib/, include/, share/ from a version directory to ~/.aether/
-    if (strcmp(sub, "--sync-from") == 0) {
-        if (argc < 2) return 1;
-        const char* ver_dir = argv[1];
-        const char* h = get_home_dir();
-        const char* subdirs[] = {"lib", "include", "share"};
-        for (int i = 0; i < 3; i++) {
-            char src_sub[1024], dest[512], cmd[4096];
-            snprintf(src_sub, sizeof(src_sub), "%s/%s", ver_dir, subdirs[i]);
-            if (dir_exists(src_sub)) {
-                snprintf(dest, sizeof(dest), "%s/.aether/%s", h, subdirs[i]);
-                snprintf(cmd, sizeof(cmd),
-                    "rm -rf \"%s\" && cp -r \"%s\" \"%s\"", dest, src_sub, dest);
-                if (system(cmd) != 0) {
-                    fprintf(stderr, "Warning: failed to sync %s to %s\n", src_sub, dest);
-                }
-            }
-        }
-        return 0;
-    }
-    // Fall-through: treat unknown sub as "ae version" (backward compat)
-    printf("ae %s (Aether Language)\n", AE_VERSION);
-    return 0;
-}
-
 
 // --------------------------------------------------------------------------
 // Cache management command
